PRESENTATION CODE
Code used to produce the analysis for the presentation. Gitea is down so I can't merge this and open a new branch so just need to continue the restructure on this branch. Going to restructure so that all the rule changes are applied individually and all at once on the same data frame